finish migrating data.rs into models and helpers, remove leftover
